rename data location paths in validator
