+++ src/app/dataValidator/helpers/validators/validateData.test.js
@@ -3,8 +3,8 @@ global.console.time = jest.fn(); // silence console.time during jest tests
 
 const { validateData } = require('./validateData');
 
-const validData = require('../../../../../data/news/test/c0000000001o.json');
-const invalidData = require('../../../../../data/news/test/c0000000023o.json');
+const validData = require('../../../../../data/test/news/c0000000001o.json');
+const invalidData = require('../../../../../data/test/news/c0000000023o.json');
 
 describe('Data Validator', () => {
   it('should not error on validateData', () => {
